Scoped down `SerDeUtils` to `ExpressionPayload.Serializer` to avoid its proliferation elsewhere; Tidying up
